cron_jobs/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
mysql-connector-python==8.0.28
protobuf==5.27.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.0
SQLAlchemy==1.4.31
urllib3==2.2.2
